+++ NEWS.md
@@ -8,10 +8,11 @@
   
   
 ## Minor changes
-   
+
+- New argument `vjust` added in `stat_pvalue_manual()`
 - `stat_pvalue_manual()` can now guess automatically the significance label column.
 - New argument `show.legend` added to `ggadd()` and `add_summary()` functions.
-
+  
 ## Bug fixes
    
 - Bug fixes in `gghistogram()`. Works now when thex variable is R keyword, such as var, mean, etc. ([#192](https://github.com/kassambara/ggpubr/issues/192))